+++ PowerPlannerAndroid/MainActivity.cs
@@ -24,6 +24,7 @@ using Android.Webkit;
 using Android.Content.PM;
 using AndroidX.Core.View;
 using Vx.Droid.Helpers;
+using PowerPlannerAppDataLibrary.App;
 
 namespace PowerPlannerAndroid
 {
@@ -94,7 +95,7 @@ namespace PowerPlannerAndroid
 
             HandleIntent();
 
-            HandleVersionChange();
+            AppUpdatedHandler.HandleAppVersionUpdated();
         }
 
 #if DEBUG
@@ -190,7 +191,7 @@ namespace PowerPlannerAndroid
             Intent = intent;
             HandleIntent();
 
-            HandleVersionChange();
+            AppUpdatedHandler.HandleAppVersionUpdated();
         }
 
         private async void HandleIntent()
@@ -328,330 +329,6 @@ namespace PowerPlannerAndroid
             {
                 TelemetryExtension.Current?.TrackEvent($"Launch_From{launchSurface}_{action}");
             }
-        }
-
-        private static bool _alreadyHandledVersionChange;
-        private void HandleVersionChange()
-        {
-            if (_alreadyHandledVersionChange)
-            {
-                // We'll make sure not to do this multiple times in single app lifecycle
-                return;
-            }
-
-            _alreadyHandledVersionChange = true;
-
-            const string VERSION = "Version";
-
-            try
-            {
-                var versionString = CrossSettings.Current.GetValueOrDefault(VERSION, null);
-
-                Version v;
-                if (versionString != null && Version.TryParse(versionString, out v))
-                {
-                }
-                else
-                {
-                    // If they have an account, that implies they were on version 5.3.16.0 which was the
-                    // last version before I started storing the version number
-                    if (AccountsManager.GetLastLoginLocalId() != Guid.Empty)
-                    {
-                        v = new Version(5, 3, 16, 0);
-                    }
-                    else
-                    {
-                        v = new Version(0, 0, 0, 0);
-                    }
-                }
-
-                // Assign new version number
-                if (v < Variables.VERSION)
-                {
-                    CrossSettings.Current.AddOrUpdateValue(VERSION, Variables.VERSION.ToString());
-                }
-
-                // If was an existing user and things have changed
-                if (v.Major > 0 && v < Variables.VERSION)
-                {
-                    string changedText = "";
-
-                    if (v <= new Version(2606, 2, 0))
-                    {
-                        changedText += "\n - Bug fix for checklists being cleared when editing a graded item";
-                    }
-
-                    if (v <= new Version(2605, 16, 0))
-                    {
-                        changedText += "\n - Fixed a bug where items on calendar sometimes disappear when moving them to different dates";
-                    }
-
-                    if (v <= new Version(2605, 10, 0, 0))
-                    {
-                        changedText += "\n - Add checklists to tasks / events! Try it when adding a task, below the details text box!";
-                    }
-
-                    if (v <= new Version(2605, 7, 0, 0))
-                    {
-                        changedText += "\n - Edit with AI beta feature! Try it on the calendar page, click the ⚡ icon.";
-                    }
-
-                    if (v <= new Version(2604, 7, 1, 99))
-                    {
-                        changedText += "\n - Custom theme colors added! Go to settings to pick your favorite color for the app!";
-                    }
-
-                    if (v <= new Version(2604, 4, 62, 99))
-                    {
-                        changedText += "\n - Fixed class reminders incorrectly appearing when there's a holiday";
-                    }
-
-                    if (v <= new Version(2603, 11, 52, 99) && v >= new Version(2601, 0))
-                    {
-                        changedText += "\n - Fixed bug where clicking calendar dates showed random days";
-                    }
-
-                    if (v <= new Version(2402, 19, 0, 0))
-                    {
-                        changedText += "\n - Ability to move semesters and classes to different semesters and years!";
-                    }
-
-                    if (v <= new Version(2303, 2, 2, 99))
-                    {
-                        changedText += "\n - Fixed issue with times appearing incorrectly on Android 14 devices";
-                    }
-
-                    if (v <= new Version(2302, 11, 2, 99))
-                    {
-                        changedText += "\n - Ability to copy semesters with their classes and schedules!\n - Ability to pick light vs dark theme!";
-                    }
-
-                    if (v <= new Version(2203, 18, 4, 0))
-                    {
-                        changedText += "\n - Improved the school time zone picker";
-                    }
-
-                    if (v <= new Version(2111, 29, 1, 99))
-                    {
-                        changedText += "\n - Added back the \"drop grade\" option!";
-                    }
-
-                    if (v <= new Version(2109, 6, 1, 99))
-                    {
-                        changedText += "\n - Ability to override final grade/GPA on years/semesters";
-                    }
-
-                    if (v <= new Version(2108, 16, 2, 99))
-                    {
-                        changedText += "\n - Ability to override final grade/GPA on classes";
-                    }
-
-                    if (v < new Version(2105, 26, 1, 99))
-                    {
-                        changedText += "\n - Class reminders will respect semester end dates (so you won't get reminders after your semester is over)";
-                    }
-
-                    if (v <= new Version(2105, 17, 1, 0))
-                    {
-                        changedText += "\n - Default grade scales! Go to the settings page to configure default grade options for all classes!";
-                    }
-
-                    if (v <= new Version(2102, 20, 1, 99))
-                    {
-                        changedText += "\n - Reminders for class schedule! See the settings to configure how soon before classes to be reminded.";
-                    }
-
-                    if (v <= new Version(2009, 23, 1, 0))
-                    {
-                        changedText += "\n - Fix for displaying calendar+day on shorter devices";
-                    }
-
-                    if (v <= new Version(2009, 17, 5, 0))
-                    {
-                        changedText += "\n - New combined calendar/day view, and Years/Settings moved to \"More\" tab on bottom right";
-                        changedText += "\n - Bug fix for list items sometimes not displaying correctly";
-                    }
-
-                    if (v <= new Version(2008, 12, 2, 0))
-                    {
-                        changedText += "\n - URLs in class schedule room fields are now clickable for supporting Zoom/online links!";
-                    }
-
-                    if (v <= new Version(2007, 30, 1, 99))
-                    {
-                        changedText += "\n - Fixed order of tasks so incomplete tasks are displayed first";
-                    }
-
-                    if (v <= new Version(2005, 30, 3, 99))
-                    {
-                        changedText += "\n - Fixes for certain time zone scenarios";
-                    }
-
-                    if (v <= new Version(2005, 22, 1, 99))
-                    {
-                        changedText += "\n - Add a grade immediately after completing a task!";
-                    }
-
-                    if (v <= new Version(2005, 19, 1, 0))
-                    {
-                        changedText += "\n - You can now convert tasks to events and vice versa!";
-                    }
-
-                    if (v <= new Version(2004, 26, 2, 0))
-                    {
-                        changedText += "\n - Dark theme support added! Power Planner will now use a dark theme based on your theme selected in Android. Even the widgets will be dark!";
-                    }
-
-                    if (v <= new Version(2003, 14, 3, 0))
-                    {
-                        changedText += "\n - Classes that don't have any grades yet will no longer count towards overall GPA";
-                    }
-
-                    if (v <= new Version(2002, 9, 1, 0))
-                    {
-                        changedText += "\n - Time zone support! If you're traveling, go to the settings page to set your school's time zone.";
-                    }
-
-                    if (v <= new Version(2001, 9, 1, 0))
-                    {
-                        changedText += "\n - Email addresses and phone numbers are now clickable in details text!";
-                    }
-
-                    if (v < new Version(1912, 23, 0))
-                    {
-                        changedText += "\n - App shortcuts added to launcher icon!";
-                    }
-
-                    if (v <= new Version(1911, 5, 2, 0))
-                    {
-                        changedText += "\n - Updated UI! Bottom navigation bar for quicker access to the most used pages.";
-                    }
-
-                    if (v < new Version(1911, 2, 0, 0))
-                    {
-                        changedText += "\n - Support for Monday as first day of week on Calendar for countries like Spain!";
-                    }
-
-                    if (v < new Version(5, 4, 86, 0))
-                    {
-                        changedText += "\n - 2x improvement in app launch speed!";
-                    }
-
-                    if (v < new Version(5, 4, 78, 0))
-                    {
-                        changedText += "\n - Image attachments on tasks/events!";
-                        changedText += "\n - Strikethrough on completed items";
-                    }
-
-                    if (v < new Version(5, 4, 72, 0))
-                    {
-                        changedText += "\n - Fixed GPA calculation handling of failed classes";
-                    }
-
-                    if (v < new Version(5, 4, 68, 0))
-                    {
-                        changedText += "\n - Pass/fail classes are now supported in the GPA grade system!\n - Custom selected due times are remembered for faster task entry";
-                    }
-
-                    if (v < new Version(5, 4, 60, 0))
-                    {
-                        changedText += "\n - Custom color picker for class colors!";
-                        changedText += "\n - 24-hour time formatting fixed on schedule view";
-                    }
-
-                    if (v < new Version(5, 4, 36, 0))
-                    {
-                        changedText += "\n - Repeating bulk entry of tasks/events!";
-                    }
-
-                    if (v < new Version(5, 4, 32, 0))
-                    {
-                        changedText += "\n - Google Calendar integration! Go to the Settings page to try it out (requires an online account)";
-                    }
-
-                    if (v < new Version(5, 4, 26, 0) && v >= new Version(5, 4, 22, 0))
-                    {
-                        // Min API is 23, so only versions below Oreo (API 26) may have had the issue
-                        if (Build.VERSION.SdkInt < BuildVersionCodes.O)
-                        {
-                            // I broke the app icon in 5.4.22.0 for people not running Oreo
-                            changedText += "\n - Fixed missing app icon";
-                        }
-                    }
-
-                    if (v < new Version(5, 4, 24, 0))
-                    {
-                        // Min API is 23 (Marshmallow), so Lollipop (API 21) APIs are always available
-                        changedText += "\n - The Android app now syncs in the background when you make a change on another device!";
-                    }
-
-                    if (v < new Version(5, 4, 22, 0))
-                    {
-                        // Show Oreo (API 26+) feature announcement
-                        if (Build.VERSION.SdkInt >= BuildVersionCodes.O)
-                        {
-                            changedText += "\n - Adaptive icon added!";
-                        }
-                    }
-
-                    if (v < new Version(5, 4, 20, 0) && v == new Version(5, 4, 18, 0))
-                    {
-                        // These only get shown to users who were running 5.4.18.0 since that's where I introduced those bugs
-                        // Min API is 23 (Marshmallow), so always show the KitKat+ message
-                        changedText += "\n - Fixed crashes when opening add grade page";
-                    }
-
-                    if (v < new Version(5, 4, 18, 0))
-                    {
-                        changedText += "\n - New Schedule widget!";
-                        changedText += "\n - Keyboard now auto-appears";
-                    }
-
-                    if (v < new Version(5, 4, 16, 0) && v >= new Version(5, 4, 12, 0))
-                    {
-                        // Only show this to Oreo+ users who were running the app targeted to Oreo (5.4.12.0)
-                        if (Build.VERSION.SdkInt >= BuildVersionCodes.O)
-                        {
-                            changedText += "\n - Fixed reminders not working on Android 8.0 Oreo devices";
-                        }
-                    }
-
-                    if (v < new Version(5, 4, 14, 0))
-                    {
-                        changedText += "\n - Pinch-to-zoom on Schedule page!";
-                    }
-
-                    if (v < new Version(5, 4, 12, 0))
-                    {
-                        changedText += "\n - Disable Autofill on everything except username/password related fields";
-                    }
-
-                    if (v < new Version(5, 4, 4, 0))
-                    {
-                        changedText += "\n - Fixed issues with automatic reminders (day-before and day-of reminders)";
-                    }
-
-                    if (v < new Version(5, 4, 0, 0))
-                    {
-                        changedText += "\n - Widget added for Agenda!";
-                    }
-
-                    if (changedText.Length > 0)
-                    {
-                        var dontWait = new PortableMessageDialog("Power Planner just installed an update!\n\nHere's what's new...\n" + changedText, "Just updated").ShowAsync();
-                    }
-                }
-            }
-
-            catch (Exception ex)
-            {
-                TelemetryExtension.Current?.TrackException(ex);
-            }
-        }
-
-        private static Guid GetLocalAccountId(QueryString qs)
-        {
-            return Guid.Parse(qs["localAccountId"]);
         }
 
         protected override void OnStop()
