+++ PowerPlannerAndroid/PowerPlannerAndroid.csproj
@@ -33,7 +33,7 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.23.0" />
-    <PackageReference Include="StorageEverywhereIncludingiOS" Version="10.0.1" />
+    <PackageReference Include="StorageEverywhereIncludingiOS" Version="10.0.4" />
     <PackageReference Include="Xam.Plugins.Settings">
       <Version>5.0.0-beta</Version>
     </PackageReference>
