+++ PowerPlannerUWP/App.xaml.cs
@@ -288,7 +288,7 @@ namespace PowerPlannerUWP
                 ConfigureDataTransferManager();
 
                 // Display updates
-                HandleVersionChange();
+                AppUpdatedHandler.HandleAppVersionUpdated();
             }
 
             catch (Exception ex)
@@ -592,276 +592,6 @@ namespace PowerPlannerUWP
         /// <returns></returns>
         private static void HandleVersionChange()
         {
-            const string VERSION = "Version";
-
-            try
-            {
-                object o;
-                Version v;
-                AppSettings.Current.TryGetValue(VERSION, out o);
-
-                // If new user
-                if (o == null)
-                {
-                    // Set to current version
-                    AppSettings.Current[VERSION] = Variables.VERSION.ToString();
-
-                    return;
-                }
-
-                v = Version.Parse(o as string);
-
-                if (v < Variables.VERSION)
-                {
-                    // Set to current version
-                    AppSettings.Current[VERSION] = Variables.VERSION.ToString();
-
-
-                    string changedText = "";
-
-                    if (v <= new Version(2606, 2, 0))
-                    {
-                        changedText += "\n - Bug fix for checklists being cleared when editing a graded item";
-                    }
-
-                    if (v <= new Version(2605, 10, 0, 0))
-                    {
-                        changedText += "\n - Add checklists to tasks / events! Try it when adding a task, below the details text box!";
-                    }
-
-                    if (v <= new Version(2605, 7, 0, 0))
-                    {
-                        changedText += "\n - Edit with AI beta feature! Try it on the calendar page, click the ⚡ icon.";
-                    }
-
-                    if (v <= new Version(2604, 7, 1, 99))
-                    {
-                        changedText += "\n - Custom theme colors added! Go to settings to pick your favorite color for the app!";
-                    }
-
-                    if (v <= new Version(2604, 4, 62, 99))
-                    {
-                        changedText += "\n - Fixed class reminders incorrectly appearing when there's a holiday";
-                    }
-
-                    if (v <= new Version(2603, 16, 58, 99) && v >= new Version(2601, 1))
-                    {
-                        changedText += "\n - Fixed 24hr time formatting issues";
-                    }
-
-                    if (v <= new Version(2603, 2, 51, 99) && v >= new Version(2601, 1))
-                    {
-                        changedText += "\n - Fixed display of class times when editing class schedule";
-                        changedText += "\n - Fixed days of week on Schedule page for users where Monday is first day of week";
-                    }
-
-                    if (v <= new Version(2408, 23, 3, 99))
-                    {
-                        changedText += "\n - Drag and drop of tasks and events is supported on nearly all the views!";
-                    }
-
-                    if (v <= new Version(2402, 19, 0, 0))
-                    {
-                        changedText += "\n - Ability to move semesters and classes to different semesters and years!";
-                    }
-
-                    if (v <= new Version(2302, 11, 2, 99))
-                    {
-                        changedText += "\n - Ability to copy semesters with their classes and schedules!\n - Ability to pick light vs dark theme!";
-                    }
-
-                    if (v <= new Version(2203, 18, 4, 0))
-                    {
-                        changedText += "\n - Improved the school time zone picker";
-                    }
-
-                    if (v <= new Version(2111, 29, 1, 99))
-                    {
-                        changedText += "\n - Added back the \"drop grade\" option!";
-                    }
-
-                    if (v <= new Version(2109, 6, 1, 99))
-                    {
-                        changedText += "\n - Ability to override final grade/GPA on years/semesters";
-                    }
-
-                    if (v <= new Version(2108, 16, 2, 99))
-                    {
-                        changedText += "\n - Ability to override final grade/GPA on classes";
-                    }
-
-                    if (v < new Version(2105, 26, 1, 99))
-                    {
-                        changedText += "\n - Class reminders will respect semester end dates (so you won't get reminders after your semester is over)";
-                    }
-
-                    if (v >= new Version(2105, 19, 3, 0) && v <= new Version(2105, 19, 3, 99))
-                    {
-                        changedText += "\n - Fixed viewing tasks/events not working";
-                    }
-
-                    if (v <= new Version(2105, 17, 1, 0))
-                    {
-                        changedText += "\n - Default grade scales! Go to the settings page to configure default grade options for all classes!";
-                    }
-
-                    if (v <= new Version(2102, 16, 3, 99))
-                    {
-                        changedText += "\n - Right click on tasks/events to quickly mark complete, delete, or edit!";
-                    }
-
-                    if (v <= new Version(2011, 25, 4, 99) && v >= new Version(2011, 25, 4, 0))
-                    {
-                        // I temporarily broke these views for one build when introducing collapsable groups
-                        changedText += "\n - Fixed \"+\" buttons on Day views";
-                    }
-
-                    if (v <= new Version(2011, 16, 1, 99))
-                    {
-                        changedText += "\n - Collapsable groups in Agenda (so you can collapse overdue items, etc)!";
-                    }
-
-                    if (v <= new Version(2008, 12, 2, 99))
-                    {
-                        changedText += "\n - Reminders for class schedule!";
-                        changedText += "\n - URLs in class schedule room fields are now clickable for supporting Zoom/online links!";
-                    }
-
-                    if (v <= new Version(2007, 30, 1, 99))
-                    {
-                        changedText += "\n - Fixed order of tasks so incomplete tasks are displayed first";
-                        changedText += "\n - Performance improvements for accounts with lots of items!";
-                    }
-
-                    if (v <= new Version(2005, 30, 3, 99))
-                    {
-                        changedText += "\n - Fixes for certain time zone scenarios";
-                    }
-
-                    if (v <= new Version(2005, 27, 1, 99))
-                    {
-                        changedText += "\n - Add a grade immediately after completing a task!";
-                    }
-
-                    if (v <= new Version(2005, 3, 1, 99))
-                    {
-                        changedText += "\n - You can now convert tasks to events and vice versa!";
-                    }
-
-                    if (v <= new Version(2005, 1, 2, 99))
-                    {
-                        changedText += "\n - Language setting added to allow overriding system selected language!";
-                        changedText += "\n - Fixed hyperlink detection bug with upper case characters";
-                    }
-
-                    if (v <= new Version(2004, 30, 1, 99))
-                    {
-                        if (v > new Version(2004, 26, 1, 99))
-                        {
-                            changedText += "\n - Back by popular demand, you can now toggle showing past complete items on the calendar view! Use the filter button in the top right.";
-                        }
-                        else
-                        {
-                            changedText += "\n - Past complete items hidden from calendar by default. You can now toggle showing past complete items by using the filter button in the top right.";
-                        }
-                    }
-
-                    if (v <= new Version(2004, 24, 2, 99))
-                    {
-                        changedText += "\n - Classes that don't have any grades yet will no longer count towards overall GPA";
-                    }
-
-                    if (v <= new Version(2003, 6, 1, 0))
-                    {
-                        changedText += "\n - Time zone support! If you're traveling, go to the settings page to set your school's time zone.";
-                    }
-
-                    if (v <= new Version(2002, 9, 1, 0))
-                    {
-                        changedText += "\n - UI fixes for tablet devices";
-                    }
-
-                    if (v <= new Version(2002, 3, 1, 0))
-                    {
-                        changedText += "\n - New text-based time pickers!";
-                    }
-
-                    if (v <= new Version(2001, 9, 1, 0))
-                    {
-                        changedText += "\n - Hyperlink detection in details text!";
-                    }
-
-                    if (v < new Version(1911, 2))
-                    {
-                        changedText += "\n - Support for Monday as first day of week on Calendar for countries like Spain!";
-                    }
-
-                    if (v < new Version(1909, 12, 1, 0))
-                    {
-                        changedText += "\n - Fixed duplicate calendar entries in Windows Calendar integration";
-                    }
-
-                    if (v < new Version(5, 4, 76, 0))
-                    {
-                        changedText += "\n - Strikethrough for completed items";
-                    }
-
-                    if (v < new Version(5, 4, 72, 0) && v >= new Version(5, 4, 64, 0))
-                    {
-                        changedText += "\n - Fixed GPA calculation for failed classes";
-                    }
-
-                    if (v < new Version(5, 4, 64, 0))
-                    {
-                        changedText += "\n - Pass/fail classes are now supported in the GPA grade system!\n - Custom selected due times are remembered for faster task entry";
-                    }
-
-                    if (v < new Version(5, 4, 60, 0) && InterfacesUWP.ColorPicker.IsCustomPickerSupported.Value)
-                    {
-                        changedText += "\n - Custom color picker for class colors!";
-                    }
-
-                    if (v < new Version(5, 4, 36, 0))
-                    {
-                        changedText += "\n - Repeating bulk entry of tasks/events!";
-                    }
-
-                    if (v < new Version(5, 4, 32, 0))
-                    {
-                        changedText += "\n - Google Calendar integration! Go to the Settings page to try it out (requires an online account)";
-                    }
-
-                    if (v < new Version(5, 4, 6, 0) && v >= new Version(5, 3, 14, 0) && !ApiInformation.IsApiContractPresent("Windows.Foundation.UniversalApiContract", 3))
-                    {
-                        // Only affects 10240 and 10586 (TH1 and TH2)
-                        changedText += "\n - Sorry for that issue with the app not launching! It was an issue that only affected older versions of Windows, and only revealed itself after being compiled for the store. Welcome back to Power Planner! :)";
-                    }
-
-                    if (v < new Version(5, 3, 14, 0))
-                    {
-                        changedText += "\n - Export schedule to image feature added to Schedule page";
-                    }
-
-                    if (v < new Version(5, 3, 12, 0))
-                    {
-                        changedText += "\n - Headers on Schedule view properly zoom out now\n - Wide schedule tile now displays more info";
-                    }
-
-                    if (v < new Version(5, 3, 0, 0))
-                    {
-                        changedText += "\n - You can now assign times to tasks/events!\n - You can add generic tasks/events that aren't under a specific class!\n - Incomplete items now appear on schedule views\n - Setting the first day of the week in two week schedule settings now updates the weekly schedule so that it starts on the same day";
-                    }
-
-
-                    if (changedText.Length > 0)
-                        MessageBox.Show("Power Planner just installed an update. Here's what's new!\n" + changedText, "Just Updated", MessageBoxButton.OK);
-                }
-            }
-
-            catch (Exception ex)
-            {
-                TelemetryExtension.Current?.TrackException(ex);
-            }
         }
 
         internal static async Task<bool> ConfirmDelete(string message, string title)
